requirements.txt
# Coral TPU Dependencies
pycoral>=2.0.0
tflite-runtime>=2.5.0
tensorflow>=2.8.0

# Data Processing
numpy>=1.21.0
pandas>=1.3.0
scikit-learn>=1.0.0

# Crypto Data APIs
ccxt>=2.0.0
yfinance>=0.1.87
requests>=2.28.0
websocket-client>=1.4.0

# Technical Analysis
ta>=0.10.2
TA-Lib>=0.4.25

# Visualization and Monitoring
matplotlib>=3.5.0
seaborn>=0.11.0
plotly>=5.0.0

# Configuration and Utilities
python-dotenv>=0.19.0
pyyaml>=6.0
schedule>=1.1.0

# Logging and Monitoring
loguru>=0.6.0

# API and Web Framework (optional)
fastapi>=0.85.0
uvicorn>=0.18.0

# Testing
pytest>=7.0.0
pytest-asyncio>=0.19.0
